Re-license everything under Apache-2.0
Much of the code and all of the intent comes from dalen-puppetdbquery
and node-puppetdbquery by Erik Dalén and is designed for use with
pypuppetdb (and possibly puppetboard) by the collective known as "Vox
Pupuli" (formerly "Puppet Community"). The four forementioned projects
are all licensed under the Apache 2.0 license, so it makes sense to make
this match to make everyone's lives easier.