Add a number of tests to increase code coverage
We're missing parts of both the lexer and parser with the current test
suite, so add three tests to make sure we cover them.

This commit exposes a deficiency in the floating pooint parsing: the
number 1.024 is parsed as 1.24 because the leading 0 in the decimal
portion is lost during the conversion to an integer.